kkn_attendance/requirements.txt
fastapi
python-multipart
uvicorn
requests
beautifulsoup4
pydantic
selenium
pillow
pytesseract
